cleaned up unused descriptions, moved log level converter
